Resolve conflict between python/ofdm/qa_bit_position_dependent_{BER,ber}.py
